refactor(build): extract remote-control server into okena-remote-server crate
Moves the HTTP/WS remote-control server (src/remote/) out of the monolithic
okena binary into a dedicated library crate, so the axum/hyper/rustls/rcgen
server stack no longer compiles on the binary's hot path and can build in
parallel. Re-exported as `crate::remote` via a shim in main.rs so existing
references keep working. CommandResult DTO moved to okena-core alongside
ActionRequest.
