+++ crates/okena-remote-server/src/routes/auth_reload.rs
@@ -1,4 +1,4 @@
-use crate::remote::routes::AppState;
+use crate::routes::AppState;
 use axum::extract::{ConnectInfo, State};
 use axum::http::StatusCode;
 use std::net::{IpAddr, SocketAddr};
